Translations done
field types updated

translating arrays not working good

+++ DevTools/Utils/MetadataManager.py
@@ -4,7 +4,7 @@ import json
 class MetadataManager:
     @staticmethod
     def get(filepath):
-        with open(filepath, 'r') as file:
+        with open(filepath, 'r', encoding='utf-8') as file:
             return json.load(file)
 
     @staticmethod
@@ -16,7 +16,7 @@ class MetadataManager:
 
     @staticmethod
     def set(section_path, key, value, filepath):
-        with open(filepath, 'r+') as file:
+        with open(filepath, 'r+', encoding='utf-8') as file:
             data = json.load(file)
             # Navigate/create the path in the data
             current_section = data
@@ -29,12 +29,35 @@ class MetadataManager:
             current_section[key] = value
 
             file.seek(0)
-            file.write(json.dumps(data, indent=4))
+            file.write(json.dumps(data, indent=4, ensure_ascii=False))
+            file.truncate()
+
+    @staticmethod
+    def set_array(section_path, index, value, filepath):
+        with open(filepath, 'r+', encoding='utf-8') as file:
+            data = json.load(file)
+            current_section = data
+            for section in section_path:
+                if section not in current_section:
+                    current_section[section] = []
+                current_section = current_section[section]
+
+            if not isinstance(current_section, list):
+                print(f"Error: Expected a list at {section_path}, found {type(current_section).__name__}.")
+                return
+
+            if len(current_section) <= index:
+                current_section.extend([None] * (index + 1 - len(current_section)))
+
+            current_section[index] = value
+
+            file.seek(0)
+            file.write(json.dumps(data, indent=4, ensure_ascii=False))
             file.truncate()
 
     @staticmethod
     def delete(section_path, key, filepath):
-        with open(filepath, 'r+') as file:
+        with open(filepath, 'r+', encoding='utf-8') as file:
             data = json.load(file)
             # Navigate to the correct section
             current_section = data
